requirements.txt
Flask==3.1.2
flask-cors==5.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.32.3
groq==0.37.1
numpy==1.26.4
scipy==1.12.0
networkx==3.2.1
scikit-fuzzy==0.4.2